# docs/swagger.yaml
definitions:
  controllers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      roleID:
        type: integer
      username:
        type: string
    required:
    - email
    - password
    - roleID
    - username
    type: object
...
  controllers.brandInput:
    properties:
      logo:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  controllers.brandUser:
    properties:
      username:
        type: string
    type: object
...
  models.Brand:
    properties:
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /brand/:
    post:
      description: Creating a new Brand (only username with role id 1 can do this).
      parameters:
      - description: the body to create a new Brand
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.brandInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
      security:
      - ApiKeyAuth: []
      summary: Create New Brand.
      tags:
      - Brand
  /brand/{id}:
    delete:
      description: Delete a Brand by id (only username with role id 1 can do this).
      parameters:
      - description: the User
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.brandUser'
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Brand.
      tags:
      - Brand
    get:
      description: Get a Brand by id.
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
      summary: Get Brand.
      tags:
      - Brand
    patch:
      description: Update Brand by id (only username with role id 1 can do this).
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update the Brand
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.brandInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Brand'
      security:
      - ApiKeyAuth: []
      summary: Update Brand.
      tags:
      - Brand
  /brand/picture/{id}:
    get:
      description: Get a Brand Picture.
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: ""
      summary: Get Brand Picture.
      tags:
      - Brand
  /brand/upload/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Upload Brand Picture (.png only!).
      parameters:
      - description: Brand id
        in: path
        name: id
        required: true
        type: string
      - description: image (.png only)
        in: formData
        name: file
        required: true
        type: file
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Upload Brand Picture.
      tags:
      - Brand
  /login:
    post:
      description: Logging in to get jwt token to access admin or Member api by roles.
      parameters:
      - description: the body to login a user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login User.
      tags:
      - Auth
...
  /register:
    post:
      description: registering a user from public access.
      parameters:
      - description: the body to register a user (Role ID 1 = Admin, Role ID 2 = Member)
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user.
      tags:
      - Auth
...
swagger: "2.0"